fix(templates): break the built-ins import cycle
vault/index -> templates-migration -> templates -> ... -> vault/index left
BUILT_IN_TEMPLATES undefined at module init, failing three suites on import.
The constants move to a leaf module both sides import directly.
